+++ R/check_model.R
@@ -47,6 +47,9 @@
 #' time-consuming. By default, `show_dots = NULL`. In this case `check_model()`
 #' tries to guess whether performance will be poor due to a very large model
 #' and thus automatically shows or hides dots.
+#' @param maximum_dots Limits the number of data points for models with many
+#' observations, to reduce the time for rendering the plot. Defaults to a
+#' maximum of 2000 data points to render
 #' @param verbose If `FALSE` (default), suppress most warning messages.
 #' @param ... Arguments passed down to the individual check functions, especially
 #' to `check_predictions()` and `binned_residuals()`.
@@ -203,6 +206,7 @@ check_model.default <- function(
   type = "density",
   residual_type = NULL,
   show_dots = NULL,
+  maximum_dots = 2000,
   size_dot = 2,
   size_line = 0.8,
   size_title = 12,
@@ -323,6 +327,7 @@ check_model.default <- function(
   attr(assumptions_data, "overdisp_type") <- list(...)$plot_type
   attr(assumptions_data, "bandwidth") <- bandwidth
   attr(assumptions_data, "type") <- type
+  attr(assumptions_data, "maximum_dots") <- maximum_dots
   attr(assumptions_data, "model_class") <- class(x)[1]
   assumptions_data
 }
@@ -358,6 +363,7 @@ check_model.stanreg <- function(
   type = "density",
   residual_type = NULL,
   show_dots = NULL,
+  maximum_dots = 2000,
   size_dot = 2,
   size_line = 0.8,
   size_title = 12,
@@ -387,6 +393,7 @@ check_model.stanreg <- function(
     bandwidth = bandwidth,
     type = type,
     residual_type = residual_type,
+    maximum_dots = maximum_dots,
     verbose = verbose,
     ...
   )
@@ -407,6 +414,7 @@ check_model.model_fit <- function(
   type = "density",
   residual_type = NULL,
   show_dots = NULL,
+  maximum_dots = 2000,
   size_dot = 2,
   size_line = 0.8,
   size_title = 12,
@@ -433,6 +441,7 @@ check_model.model_fit <- function(
     base_size = base_size,
     detrend = detrend,
     show_dots = show_dots,
+    maximum_dots = maximum_dots,
     bandwidth = bandwidth,
     type = type,
     residual_type = residual_type,
@@ -452,6 +461,7 @@ check_model.performance_simres <- function(
   type = "density",
   residual_type = NULL,
   show_dots = NULL,
+  maximum_dots = 2000,
   size_dot = 2,
   size_line = 0.8,
   size_title = 12,
@@ -478,6 +488,7 @@ check_model.performance_simres <- function(
     base_size = base_size,
     detrend = detrend,
     show_dots = show_dots,
+    maximum_dots = maximum_dots,
     bandwidth = bandwidth,
     type = type,
     residual_type = "simulated",